+++ Grid_Tab.pde
@@ -22,17 +22,17 @@ public class Grid {
     for (int i=0; i<15; i++) { // 15 columns
       for (int j=0; j<17; j++) { // 17 rows
         if (i % 2 == 0){
-          this.cellGrid[i][j] = new Cell(new Bubble((i<9) ? aColour.randomColour() : INV,false), int(gAPO+2*gAPO*j), int(gRAD+1.5*gRAD*i));
+          this.cellGrid[i][j] = new Cell(new Bubble((i<9) ? aColour.randomColour() : INV,false), gAPO+2*gAPO*j, gRAD+1.5*gRAD*i);
           //System.out.println(this.cellGrid[i][j].bubble.col);
         }
         else
-          this.cellGrid[i][j] = new Cell(new Bubble((i<9) ? aColour.randomColour() : INV,false), int(2*gAPO+(2*gAPO*j)), int(gRAD+1.5*gRAD*i));
+          this.cellGrid[i][j] = new Cell(new Bubble((i<9) ? aColour.randomColour() : INV,false), 2*gAPO+(2*gAPO*j), gRAD+1.5*gRAD*i);
       }
     }
     for (int i=0; i<6; i++) {
-      this.bottomCellGrid[i] = new Cell(new Bubble(INV,true), int(2*gAPO+37*i), int(gRAD+1.5*gRAD*16));
+      this.bottomCellGrid[i] = new Cell(new Bubble(INV,true), 2*gAPO+37*i, gRAD+1.5*gRAD*16);
     }
-    this.mainCell = new Cell(new Bubble(aColour.randomColour(), false), int(gAPO+2*gAPO*8), int(gRAD+1.5*gRAD*16));
+    this.mainCell = new Cell(new Bubble(aColour.randomColour(), false), gAPO+2*gAPO*8, gRAD+1.5*gRAD*16);
     this.bottomCellGrid[0].bubble.col=aColour.randomColour();
   }
